[VXXX] Refactor and apply further code review suggestions
